check plugin md5

+++ scripts/setup-ci-environment.sh
@@ -2,21 +2,13 @@
 
 source scripts/prepare-ci-environment.sh
 
-# we will use Community ids to download plugins.
-SCALA_PLUGIN_ID="org.intellij.scala"
-PYTHON_PLUGIN_ID="PythonCore"
-if [[ $IJ_ULTIMATE == "true" ]]; then
-  PYTHON_PLUGIN_ID="Pythonid"
-fi
-
 mkdir -p .cache/intellij/$FULL_IJ_BUILD_NUMBER
 
 if [ ! -d .cache/intellij/$FULL_IJ_BUILD_NUMBER/idea-dist ]; then
   IJ_TAR_NAME=idea${IJ_BUILD}.tar.gz
   echo "Loading $IJ_BUILD..."
   wget -O $IJ_TAR_NAME http://download.jetbrains.com/idea/idea${IJ_BUILD}.tar.gz
-  tar_md5=$(md5sum $IJ_TAR_NAME  | awk -F " " '{print $1}')
-  if [ $tar_md5 != $EXPECTED_IJ_MD5 ];
+  if [ $(md5sum $IJ_TAR_NAME  | awk -F " " '{print $1}') != $EXPECTED_IJ_MD5 ];
   then
     echo "IJ tar md5 incorrect" >&2
     exit 1
@@ -38,11 +30,22 @@ if [ ! -d .cache/intellij/$FULL_IJ_BUILD_NUMBER/plugins ]; then
   pushd plugins
 
   wget -O Scala.zip "https://plugins.jetbrains.com/pluginManager/?action=download&id=$SCALA_PLUGIN_ID&build=$FULL_IJ_BUILD_NUMBER"
+  if [ $(md5sum Scala.zip  | awk -F " " '{print $1}') != $SCALA_PLUGIN_MD5 ];
+  then
+    echo "Scala plugin md5 incorrect" >&2
+    exit 1
+  fi
   unzip Scala.zip
-  #rm -rf Scala.zip
+  rm -rf Scala.zip
+
   wget -O python.zip "https://plugins.jetbrains.com/pluginManager/?action=download&id=$PYTHON_PLUGIN_ID&build=$FULL_IJ_BUILD_NUMBER"
+  if [ $(md5sum python.zip  | awk -F " " '{print $1}') != $PYTHON_PLUGIN_MD5 ];
+  then
+    echo "Scala plugin md5 incorrect" >&2
+    exit 1
+  fi
   unzip python.zip
-  #rm -rf python.zip
+  rm -rf python.zip
 
   popd
   mv plugins ".cache/intellij/$FULL_IJ_BUILD_NUMBER/plugins"
